+++ src/lib.rs
@@ -7,7 +7,6 @@ mod app;
 mod color;
 mod console;
 mod file;
-//mod font;
 mod img;
 mod input;