+++ examples/basic.rs
@@ -2,6 +2,16 @@ extern crate doryen_rs;
 
 use doryen_rs::{App, AppOptions, DoryenApi, Engine, TextAlign, UpdateEvent};
 
+// this part makes it possible to compile to wasm32 target
+#[cfg(target_arch = "wasm32")]
+use wasm_bindgen::prelude::*;
+#[cfg(target_arch = "wasm32")]
+#[wasm_bindgen(start)]
+pub fn main_js() -> Result<(), JsValue> {
+    main();
+    Ok(())
+}
+
 /*
 Apart from the basic real-time walking, this example shows how screenshots can be captured in-game.
 Because it uses UpdateEvent, any combination of keys can be specified to activate it.