+++ www/index.js
@@ -1,0 +1,2 @@
+import('../examples/pkg')
+    .catch(console.error);