+++ CHANGELOG.md
@@ -1,7 +1,8 @@
 # 2.0.0 - TBD
 * switched from unrust to bracket-lib
 ## API breaks
-* fonts path are defined in `AppOptions::font_paths`. `DoryenApi::set_font_path` replaced with `DoryenApi::set_font_index`
+* fonts path are defined in `AppOptions.font_paths`. `DoryenApi.set_font_path` replaced with `DoryenApi.set_font_index`
+* On native target, the window is always resizeable. If `AppOptions.resizeable` is true, the console is resized with the window, else it's scaled (default behavior). You don't have to call `Console.resize`anymore (see resize example).
 
 # 1.2.4 - TBD
 ## features